+++ far/far2sdk/farplug-wide.h
@@ -592,7 +592,7 @@ struct OpenDlgPluginData
 
 struct DialogInfo
 {
-	int StructSize;
+	size_t StructSize;
 	GUID Id;
 	DWORD Owner;
 };
@@ -1740,7 +1740,7 @@ struct ViewerMode
 
 struct ViewerInfo
 {
-	int    StructSize;
+	size_t StructSize;
 	int    ViewerID;
 	const wchar_t *FileName;
 	int64_t FileSize;
@@ -2435,7 +2435,7 @@ enum BOX_DEF_SYMBOLS
 
 typedef struct FarStandardFunctions
 {
-	int StructSize;
+	size_t StructSize;
 
 	FARSTDATOI                 atoi;
 	FARSTDATOI64               atoi64;
@@ -2529,7 +2529,7 @@ typedef struct FarStandardFunctions
 
 struct PluginStartupInfo
 {
-	int StructSize;
+	size_t StructSize;
 	const wchar_t *ModuleName;
 	INT_PTR ModuleNumber;
 	const wchar_t *RootKey;
@@ -2599,7 +2599,7 @@ enum PLUGIN_FLAGS
 
 struct PluginInfo
 {
-	int StructSize;
+	size_t StructSize;
 	DWORD Flags;
 	const wchar_t * const *DiskMenuStrings;
 	int *Reserved0;
@@ -2734,7 +2734,7 @@ FAR_INLINE_CONSTANT OPERATION_MODES
 
 struct OpenPluginInfo
 {
-	int                           StructSize;
+	size_t                        StructSize;
 	DWORD                         Flags;
 	const wchar_t*                HostFile;
 	const wchar_t*                CurDir;
