+++ far/src/panels/filelist.cpp
@@ -4630,7 +4630,7 @@ bool FileList::UpdateKeyBar()
 			KB->SetAlt((const wchar_t **)Info.KeyBar->AltTitles, 12);
 			KB->SetCtrl((const wchar_t **)Info.KeyBar->CtrlTitles, 12);
 
-			if ((size_t)Info.StructSize >= sizeof(OpenPluginInfo)) {
+			if (Info.StructSize >= sizeof(OpenPluginInfo)) {
 				KB->SetCtrlShift((const wchar_t **)Info.KeyBar->CtrlShiftTitles, 12);
 				KB->SetAltShift((const wchar_t **)Info.KeyBar->AltShiftTitles, 12);
 				KB->SetCtrlAlt((const wchar_t **)Info.KeyBar->CtrlAltTitles, 12);
