+++ far/src/plug/PluginA.cpp
@@ -1055,7 +1055,7 @@ void PluginA::ConvertOpenPluginInfo(oldfar::OpenPluginInfo &Src, OpenPluginInfo 
 	if (Src.KeyBar)
 	{
 		OPI.KeyBar=(KeyBarTitles*) malloc(sizeof(KeyBarTitles));
-		ConvertKeyBarTitlesA(Src.KeyBar, (KeyBarTitles*)OPI.KeyBar, Src.StructSize>=(int)sizeof(oldfar::OpenPluginInfo));
+		ConvertKeyBarTitlesA(Src.KeyBar, (KeyBarTitles*)OPI.KeyBar, Src.StructSize>=sizeof(oldfar::OpenPluginInfo));
 	}
 
 	if (Src.ShortcutData)
