+++ luafar/lua_share/far2/farapi.lua
@@ -437,7 +437,7 @@ struct OpenDlgPluginData
 
 struct DialogInfo
 {
-	int StructSize;
+	size_t StructSize;
 	GUID Id;
 	DWORD Owner;
 };
@@ -1497,7 +1497,7 @@ struct ViewerMode
 
 struct ViewerInfo
 {
-	int    StructSize;
+	size_t StructSize;
 	int    ViewerID;
 	const wchar_t *FileName;
 	int64_t FileSize;
@@ -2145,7 +2145,7 @@ enum BOX_DEF_SYMBOLS
 
 typedef struct FarStandardFunctions
 {
-	int StructSize;
+	size_t StructSize;
 
 	FARSTDATOI                 atoi;
 	FARSTDATOI64               atoi64;
@@ -2235,7 +2235,7 @@ typedef struct FarStandardFunctions
 
 struct PluginStartupInfo
 {
-	int StructSize;
+	size_t StructSize;
 	const wchar_t *ModuleName;
 	INT_PTR ModuleNumber;
 	const wchar_t *RootKey;
@@ -2305,7 +2305,7 @@ enum PLUGIN_FLAGS
 
 struct PluginInfo
 {
-	int StructSize;
+	size_t StructSize;
 	DWORD Flags;
 	const wchar_t * const *DiskMenuStrings;
 	int *Reserved0;
@@ -2431,7 +2431,7 @@ enum {
 
 struct OpenPluginInfo
 {
-	int                           StructSize;
+	size_t                        StructSize;
 	DWORD                         Flags;
 	const wchar_t*                HostFile;
 	const wchar_t*                CurDir;
